Changed code to export set_operating_point.
I changed the code to export the set_operating_point method.
This allows for quickly changing the operating point without
having to recreate the simulation object. This is useful for
recalculating the jacobian of the system after having changed
some system parameter such as generator active set power. I also
moved the function to calculate the reduced jacobian from the
small_signal.jl file to the jacobian.jl file.